# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateLinkReq:
    properties:
      original_url:
        type: string
      slug:
        type: string
    required:
    - original_url
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      isSucces:
        type: boolean
      message:
        type: string
    type: object
  dto.LinkDetailRes:
    properties:
      click_count:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      original_url:
        type: string
      short_url:
        type: string
      slug:
        type: string
    type: object
  dto.LoginReq:
    properties:
      email:
        example: koda36@gmail.com
        type: string
      password:
        example: Password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.LoginUserDetail'
    type: object
  dto.LoginUserDetail:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
  dto.RegisterReq:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    type: object
  dto.RegisterRes:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  dto.Response-any:
    properties:
      data: {}
      isSucces:
        type: boolean
      message:
        type: string
    type: object
  dto.Response-array_dto_LinkDetailRes:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.LinkDetailRes'
        type: array
      isSucces:
        type: boolean
      message:
        type: string
    type: object
  dto.Response-dto_LinkDetailRes:
    properties:
      data:
        $ref: '#/definitions/dto.LinkDetailRes'
      isSucces:
        type: boolean
      message:
        type: string
    type: object
  dto.Response-dto_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      isSucces:
        type: boolean
      message:
        type: string
    type: object
  dto.Response-dto_RegisterRes:
    properties:
      data:
        $ref: '#/definitions/dto.RegisterRes'
      isSucces:
        type: boolean
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend shortlink created by Hanif using Gin
  license:
    name: MIT
  title: Backend Shortlink
  version: "1.0"
paths:
  /api/links:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all active short links created by the logged-in
        user.
      produces:
      - application/json
      responses:
        "200":
          description: List of user shortlinks retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_LinkDetailRes'
        "401":
          description: Unauthorized - Token missing or invalid
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error - Database or context error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all short links for the authenticated user
      tags:
      - Link Management
    post:
      consumes:
      - application/json
      description: Authenticated users can transform a long destination URL into a
        clean, short link. If custom slug is omitted, the system will auto-generate
        a random unique identifier.
      parameters:
      - description: Create Link Payload (destination URL and optional custom slug)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLinkReq'
      produces:
      - application/json
      responses:
        "201":
          description: Shortlink created successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_LinkDetailRes'
        "400":
          description: Bad Request - Validation error, reserved keyword used, or duplicate
            slug
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized - Token missing or invalid signature
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error - Invalid token payload context format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new short link
      tags:
      - Link Management
  /api/links/{id}:
    delete:
      consumes:
      - application/json
      description: Authenticated users can soft delete their own short links. The
        cache in Redis will be invalidated instantly.
      parameters:
      - description: Short Link ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Shortlink deleted successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized - Token invalid or missing
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "410":
          description: Gone/Not Found - Link not found or already deleted
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Soft delete a short link
      tags:
      - Link Management
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password to get a JWT token
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.Response-dto_LoginResponse'
        "400":
          description: Invalid input validation
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized - Invalid email or password
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /api/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/dto.Response-dto_RegisterRes'
        "400":
          description: Invalid input validation/ password mismatch
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer token used for authorization (Format: Bearer <token>)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"